<Dropdown/> - remove deprecared uncontrolled mode
