+++ src/Dropdown/Dropdown.js
@@ -1,10 +1,7 @@
 import classNames from 'classnames';
-import PropTypes from 'prop-types';
-import isUndefined from 'lodash/isUndefined';
 import defaultTo from 'lodash/defaultTo';
 import differenceBy from 'lodash/differenceBy';
 import { allValidators, extendPropTypes } from '../utils/propTypes';
-import deprecationLog from '../utils/deprecationLog';
 
 import InputWithOptions from '../InputWithOptions/InputWithOptions';
 import styles from './Dropdown.scss';
@@ -14,27 +11,20 @@ const NO_SELECTED_ID = null;
 class Dropdown extends InputWithOptions {
   constructor(props) {
     super(props);
-    if (props.upgrade) {
-      this.state = {
-        value: '',
-        selectedId: NO_SELECTED_ID,
-        ...Dropdown.getNextState(
-          props,
-          defaultTo(props.selectedId, props.initialSelectedId),
-        ),
-      };
-    } else {
-      this.deprecatedUpdate(props, { isFirstTime: true });
-    }
+
+    this.state = {
+      value: '',
+      selectedId: NO_SELECTED_ID,
+
+      ...Dropdown.getNextState(
+        props,
+        defaultTo(props.selectedId, props.initialSelectedId),
+      ),
+    };
   }
 
   isSelectedIdControlled() {
-    const { upgrade, selectedId } = this.props;
-    return upgrade && !isUndefined(selectedId);
-  }
-
-  isControlledSupported() {
-    return this.props.upgrade;
+    return typeof this.props.selectedId !== 'undefined';
   }
 
   static isOptionsEqual(optionsA, optionsB) {
@@ -42,14 +32,11 @@ class Dropdown extends InputWithOptions {
   }
 
   getSelectedId() {
-    if (this.isControlledSupported()) {
-      return this.isSelectedIdControlled()
-        ? this.props.selectedId
-        : this.state.selectedId;
-    } else {
-      return this.state.selectedId;
-    }
+    return this.isSelectedIdControlled()
+      ? this.props.selectedId
+      : this.state.selectedId;
   }
+
   _onInputClicked(event) {
     if (
       this.state.showOptions &&
@@ -70,7 +57,7 @@ class Dropdown extends InputWithOptions {
    * If selectedId is not found in options, then value is NOT changed.
    */
   static getNextState(props, selectedId) {
-    if (!isUndefined(selectedId)) {
+    if (typeof selectedId !== 'undefined') {
       const option = props.options.find(_option => {
         return _option.id === selectedId;
       });
@@ -85,42 +72,17 @@ class Dropdown extends InputWithOptions {
     return {};
   }
 
-  deprecatedUpdate(props, { isFirstTime }) {
-    let value = '',
-      selectedId = -1;
-    if (!isUndefined(props.selectedId)) {
-      const option = props.options.find(_option => {
-        return _option.id === props.selectedId;
-      });
-
-      if (option) {
-        value = props.valueParser(option);
-        selectedId = option.id;
-      }
-    }
-
-    if (isFirstTime) {
-      this.state = { value, selectedId };
-    } else {
-      this.setState({ value, selectedId });
-    }
-  }
-
   componentWillReceiveProps(nextProps) {
-    if (this.props.upgrade) {
-      if (
-        nextProps.selectedId !== this.props.selectedId ||
-        !Dropdown.isOptionsEqual(this.props.options, nextProps.options)
-      ) {
-        this.setState(
-          Dropdown.getNextState(
-            nextProps,
-            defaultTo(nextProps.selectedId, this.state.selectedId),
-          ),
-        );
-      }
-    } else {
-      this.deprecatedUpdate(nextProps, { isFirstTime: false });
+    if (
+      nextProps.selectedId !== this.props.selectedId ||
+      !Dropdown.isOptionsEqual(this.props.options, nextProps.options)
+    ) {
+      this.setState(
+        Dropdown.getNextState(
+          nextProps,
+          defaultTo(nextProps.selectedId, this.state.selectedId),
+        ),
+      );
     }
   }
 
@@ -146,7 +108,7 @@ class Dropdown extends InputWithOptions {
   }
 
   _onSelect(option) {
-    if (!this.isControlledSupported() || !this.isSelectedIdControlled()) {
+    if (!this.isSelectedIdControlled()) {
       this.setState({
         value: this.props.valueParser(option),
         selectedId: option.id,
@@ -163,8 +125,6 @@ class Dropdown extends InputWithOptions {
 
 Dropdown.propTypes = {
   ...InputWithOptions.propTypes,
-  /** When true, then `selectedId` is used for Controlled mode, and `initialSelectedId` for Uncontrolled mode. */
-  upgrade: PropTypes.bool,
   /** Sets the selected option id. (Implies Controlled mode) */
   selectedId: InputWithOptions.propTypes.selectedId,
   /** An initial selected option id. (Implies Uncontrolled mode) */
@@ -185,24 +145,8 @@ extendPropTypes(Dropdown, {
       }
     },
   ),
-  initialSelectedId: allValidators(
-    InputWithOptions.propTypes.selectedId,
-    (props, propName) => {
-      if (props[propName] !== undefined && !props['upgrade']) {
-        return new Error(
-          `'initialSelectedId' can be used only if you pass 'upgrade=true' as well.`,
-        );
-      }
-    },
-  ),
-  upgrade: allValidators(PropTypes.bool, (props, propName) => {
-    if (!props[propName]) {
-      deprecationLog(
-        `Dropdown: New API! Please upgrade by passing the prop 'upgrade=true', and refer to documentation.`,
-      );
-    }
-  }),
 });
+
 Dropdown.defaultProps = InputWithOptions.defaultProps;
 Dropdown.displayName = 'Dropdown';
